+++ lib/rspec/mocks/mutate_const.rb
@@ -399,11 +399,5 @@ module RSpec
         end
       end
     end
-
-    # Keeps backwards compatibility since we had released an rspec-mocks that
-    # only supported stubbing. Later, we released the hide_const feature and
-    # decided that the term "mutator" was a better term to wrap up the concept
-    # of both stubbing and hiding.
-    ConstantStubber = ConstantMutator
   end
 end