+++ lib/rspec/mocks/verifying_proxy.rb
@@ -114,6 +114,7 @@ module RSpec
       end
 
       def add_expectation(*args, &block)
+        # explict params necessary for 1.8.7 see #626
         super(*args, &block).tap { |x| x.method_reference = @method_reference }
       end