+++ Changelog.md
@@ -4,6 +4,7 @@
 Deprecations:
 
 * Deprecate `RSpec::Mocks::TestDouble.extend_onto`. (Myron Marston)
+* Deprecate `RSpec::Mocks::ConstantStubber`. (Jon Rowe)
 
 ### 2.99.0.beta2 / 2014-02-17
 [full changelog](http://github.com/rspec/rspec-mocks/compare/v2.99.0.beta1...v2.99.0.beta2)