+++ Changelog.md
@@ -31,6 +31,8 @@ Bug Fixes:
   (Maurício Linhares)
 * Prevent `StackLevelTooDeep` errors when stubbing an `any_instance` method that's
   accessed in `inspect` by providing our own inspect output. (Jon Rowe)
+* Fix bug in `any_instance` logic that dd not allow you to mock or stub
+  private methods if `verify_partial_doubles` was configured. (Oren Dobzinski)
 
 ### 3.0.0.beta2 / 2014-02-17
 [Full Changelog](http://github.com/rspec/rspec-mocks/compare/v3.0.0.beta1...v3.0.0.beta2)